Move back macros in main package.
