+++ buildSrc/src/main/java/io/micronaut/guides/core/RawTestMacroSubstitution.java
@@ -1,8 +1,5 @@
-package io.micronaut.guides.core.macros;
+package io.micronaut.guides.core;
 
-import io.micronaut.guides.core.GuidesConfiguration;
-import io.micronaut.guides.core.GuidesOption;
-import io.micronaut.guides.core.LicenseLoader;
 import jakarta.inject.Singleton;
 
 @Singleton
