+++ buildSrc/src/test/java/io/micronaut/guides/core/RawTestMacroSubstitutionTest.java
@@ -1,6 +1,5 @@
-package io.micronaut.guides.core.macros;
+package io.micronaut.guides.core;
 
-import io.micronaut.guides.core.GuidesOption;
 import io.micronaut.starter.api.TestFramework;
 import io.micronaut.starter.options.BuildTool;
 import io.micronaut.starter.options.Language;
